添加Action及ActionComponent

+++ Source/ActionRoguelike/Public/SPowerup_HealthPotion.h
@@ -21,7 +21,7 @@ public:
 protected:
 
 	UPROPERTY(EditAnywhere, Category="Powerup")
-	float CreditCost;
+	int CreditCost;
 
 	UPROPERTY(EditAnywhere, Category="Powerup")
 	float HealthHeal;
